django_project_final/fundraising_project/requirements.txt
Django==5.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1
drf-yasg==1.21.7
Pillow==12.1.0
python-dotenv==1.2.1